+++ Assets/CommonCore/CoreShared/Util/TextureAssignScript.cs
@@ -1,11 +1,8 @@
-﻿using CommonCore.World;
-using System;
-using System.Collections;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 
-namespace CommonCore.Experimental
+namespace CommonCore.Util
 {
 
     /// <summary>