+++ Assets/CommonCoreModules/World/ObjectComponents/FacingSpriteAssignScript.cs
@@ -1,10 +1,9 @@
 ﻿using CommonCore.World;
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace CommonCore.Experimental
+namespace CommonCore.World
 {
 
     /// <summary>